+++ lib/polyfill/v2_2.rb
@@ -1,5 +1,6 @@
 require_relative 'v2_2/enumerable'
 require_relative 'v2_2/kernel'
+require_relative 'v2_2/vector'
 
 module Polyfill
   module V2_2