+++ spec/polyfill/v2_2/vector/instance/plus_unary_spec.rb
@@ -1,0 +1,10 @@
+require 'matrix'
+
+RSpec.describe 'Vector#+@' do
+  using Polyfill(Vector: %w[#+@])
+
+  it 'returns self' do
+    obj = Vector[]
+    expect(+obj).to be obj
+  end
+end