+++ CHANGELOG.md
@@ -8,6 +8,10 @@
  - v2.2 Enumerable#min_by
  - Polyfill.get can now be used with `prepend`
 
+## Fixed
+
+ - Enumerable should add to Matrix and Vector as well
+
 # [0.8.0][] (2017-04-27)
 
 ## Changed